refactor: atomic SharedDemoTime.tick, shared tickDemoFrame, fix bundle prefix
- Replace separate syncPause/elapsed/angle methods with single tick()
  that acquires the Mutex once for atomic pause+elapsed+angle computation
- Extract tickDemoFrame() shared by both render delegates, eliminating
  code duplication and adding FPS dispatch to the Native tab
- Simplify both drawInMTKView to one-liner calling tickDemoFrame()
- Remove unnecessary remember { sharedDemoStore } in ComposeIosEntry
- Add KDoc on IosDemoHandle.renderDelegate explaining GC retention
- Fix bundleIdPrefix to com.hyeonslab.prism.ios matching bundle ID
