+++ devlog/000008-feat-sync-ios-tab-rotation.md
@@ -80,15 +80,18 @@ Rename the `prism-ios-demo/` Xcode project from `PrismDemo`/`PrismDemoApp` to `P
 - **[2026-02-15 PST]** `prism-ios-demo/project.yml` — Renamed project `PrismDemo` → `PrismiOSDemo`, target `PrismDemoApp` → `PrismiOSDemo`, product name `PrismDemoApp` → `PrismiOSDemo`, bundle ID `com.hyeonslab.prism.demo` → `com.hyeonslab.prism.ios.demo`.
 - **[2026-02-15 PST]** `prism-demo/README.md` — Updated xcodebuild scheme `PrismDemoApp` → `PrismiOSDemo`, xcodeproj `PrismDemo.xcodeproj` → `PrismiOSDemo.xcodeproj`, app name `PrismDemoApp.app` → `PrismiOSDemo.app`, bundle ID `com.hyeonslab.prism.demo` → `com.hyeonslab.prism.ios.demo`. Added note explaining the relationship between `prism-demo` (KMP module) and `prism-ios-demo` (Swift consumer app).
 - **[2026-02-15 PST]** `prism-ios-demo/README.md` — New file. Documents the iOS demo app, its relationship to the KMP `prism-demo` module, build/run instructions, and project structure.
-- **[2026-02-15 PST]** `IosConstants.kt` — Added `Mutex`-based synchronization to `SharedDemoTime` for thread safety across display-link threads. When the lock is contended, cached values from the previous frame are returned (no action runs without the lock). Callers are synchronous ObjC callbacks (`drawInMTKView`) so suspend is not viable.
-- **[2026-02-15 PST]** `IosDemoController.kt`, `ComposeIosEntry.kt` — Material allocation now guarded by color-change check to avoid per-frame GC pressure.
+- **[2026-02-15 PST]** `IosConstants.kt` — Replaced separate `syncPause()`/`elapsed()`/`angle()` methods with a single `tick()` inline method that acquires the Mutex once per frame for atomic pause+elapsed+angle computation. Added `tickDemoFrame()` shared function that both render delegates call, deduplicating rotation, material, FPS, and ECS update logic.
+- **[2026-02-15 PST]** `IosDemoController.kt` — Simplified `DemoRenderDelegate.drawInMTKView()` to call shared `tickDemoFrame()`. Added KDoc on `renderDelegate` parameter explaining why it exists (prevents GC). Removed 7 unused imports.
+- **[2026-02-15 PST]** `ComposeIosEntry.kt` — Simplified `ComposeRenderDelegate.drawInMTKView()` to call shared `tickDemoFrame()`. Removed unnecessary `remember { sharedDemoStore }` (top-level val doesn't need remember). Removed 8 unused imports.
 - **[2026-02-15 PST]** `ComposeDemoControls.kt` — Added `contentDescription` accessibility labels to color swatch buttons.
+- **[2026-02-15 PST]** `prism-ios-demo/project.yml` — Fixed `bundleIdPrefix` from `com.hyeonslab.prism` to `com.hyeonslab.prism.ios` to match the explicit `PRODUCT_BUNDLE_IDENTIFIER`.
 
 ### Decisions
 - **[2026-02-15 PST]** **`PrismiOSDemo` as new name** — Clear disambiguation: `PrismDemo` is the KMP XCFramework module name (used by `import PrismDemo` in Swift), while `PrismiOSDemo` is the iOS app project/target/scheme name.
 - **[2026-02-15 PST]** **New bundle ID `com.hyeonslab.prism.ios.demo`** — Adds `.ios` segment to distinguish from any future demo apps on other platforms.
 - **[2026-02-15 PST]** **Swift `import PrismDemo` unchanged** — This imports the KMP XCFramework module, not the app. No Swift source changes needed.
-- **[2026-02-15 PST]** **Mutex over @Volatile or suspend** — `drawInMTKView` is a synchronous ObjC protocol callback from Metal's display link — it cannot be `suspend`. `@Volatile` gives visibility but not atomicity across multi-field reads/writes. `Mutex.tryLock` is non-blocking: if contended, cached values are returned instead of running the action without the lock. K/N new memory model guarantees cross-thread visibility for object property writes, so the cached fields don't need `@Volatile`.
+- **[2026-02-15 PST]** **Mutex with single `tick()` over separate methods** — `drawInMTKView` is a synchronous ObjC callback — cannot use `suspend`. Replaced separate `syncPause()`/`elapsed()`/`angle()` (each acquiring the lock independently, allowing inconsistent reads between calls) with a single `tick()` that acquires once and computes all values atomically. Cached fields for contention fallback don't use `@Volatile` — at 60fps a single frame of staleness is imperceptible.
+- **[2026-02-15 PST]** **Shared `tickDemoFrame()` function** — Eliminated code duplication between `DemoRenderDelegate` and `ComposeRenderDelegate`. Both delegates now call `tickDemoFrame()` which handles rotation, material color, FPS dispatch, and ECS update. This also fixes the Native tab not dispatching FPS updates.
 - **[2026-02-15 PST]** **Generic simulator destination** — Replaced `iPhone 16 Pro` with `generic/platform=iOS Simulator` in both READMEs for broader Xcode version compatibility.
 
 ### Commits
